feat: implemented join_game

+++ poker-texas-hold-em/contract/src/models/game.cairo
@@ -18,7 +18,7 @@ pub enum GameMode {
 #[derive(Copy, Drop, Serde, Default, Introspect, PartialEq)]
 pub struct GameParams {
     game_mode: GameMode,
-    max_no_of_players: u8,
+    max_no_of_players: u32,
     small_blind: u64,
     big_blind: u64,
     no_of_decks: u8,
